table booking demo mvc

+++ springBasics/restaurantBookingApplication/restaurant.booking/src/main/java/com/example/restaurant/booking/repository/TableRepository.java
@@ -3,6 +3,6 @@ package com.example.restaurant.booking.repository;
 import com.example.restaurant.booking.model.Table;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TableRepository extends JpaRepository<Table,Long> {
+public interface TableRepository extends JpaRepository<Table, Integer> {
 
 }
